fix(golangci_lint_ls): disable all output formats except JSON
This change set enables all possible output formats that exist for
golangci-lint. This is necessary because if a user configures output
formats in their `.golangci.yml` configuration (for example for CI) then
this is always in addition to the ones enabled on the command line.
Therefore, we manually disable all possible output formats.

Currently, there doesn't exist a `golangci-lint` option to exclusively
enable an output format or disable all from the config.